+++ src/edcon_tools/generic_bus_argparser.py
@@ -1,5 +1,6 @@
 import argparse
 import logging
+from rich.logging import RichHandler
 
 
 class GenericBusArgParser():
@@ -38,5 +39,8 @@ class GenericBusArgParser():
         args = parser.parse_args()
 
         if args.verbose:
-            logging.basicConfig(format='%(message)s', level=logging.INFO)
+            logging.basicConfig(level=logging.INFO,
+                                format='%(message)s',
+                                datefmt="[%X]",
+                                handlers=[RichHandler()])
         return args